requirements_api.txt
starlette_exporter
sqlalchemy
pydantic
psycopg2
python-slugify
pyjwt
pandas
fastapi>=0.70.0
requests

